verba/requirements.txt
weaviate-client==3.25.2
openai==1.3.5
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-openai==0.0.8
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.1
tiktoken==0.5.1
